Refactoring theprofile page to display dynamically
