+++ client/src/components/Sections/ShowProfile.js
@@ -1,0 +1,9 @@
+import React from "react"
+
+export const ShowProfile = () => {
+  return (
+    <>
+      <div>My Profile</div>
+    </>
+  )
+}
